// src/app/models/FinalResponse.ts
export class FinalResponse {
  claimId?: string;
      ProviderLastName?: string;
      ProviderNPI?: string;
      ProviderAddress1?: string;
      CityName?: string;
      ZipCode?: string;
      GenderCode?: string;
      PatAcct?: string;
      HeaderChargeAmt?: string;
      AdmitDate?: string;
      SubmittypeCode?: string;
      IsCurrent?: string;
            PaidAmount?: string;
      IsCapitated?: string;
      IsScrubbed?: string;
      ScoredLables?: string;
      ScoredProbabilities?: string;
}